refactor users-places routes
