+++ src/places/controllers/places.controller.spec.ts
@@ -3,11 +3,11 @@ import { Sequelize } from 'sequelize-typescript';
 import { getModelToken } from '@nestjs/sequelize';
 
 import { PlacesController } from './places.controller';
-import { PlacesService } from './services/places.service';
-import { UsersService } from '../users/users.service';
-import { User } from '../users/models/user.model';
-import { DemandsService } from '../demands/services/demands.service';
-import { JwtService } from '../jwt/jwt.service';
+import { PlacesService } from '../services/places.service';
+import { UsersService } from '../../users/users.service';
+import { User } from '../../users/models/user.model';
+import { DemandsService } from '../../demands/services/demands.service';
+import { JwtService } from '../../jwt/jwt.service';
 
 describe('PlacesController', () => {
   let controller: PlacesController;
